2024.12.13

+++ src/main/java/wizard/eVC/infoMgmt/infoAdd/InfoAddDetailController.java
@@ -12,14 +12,15 @@ package wizard.eVC.infoMgmt.infoAdd;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
-import wizard.eVC.common.util.FTP;
+
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 import wizard.eVC.infoMgmt.infoAdd.dto.InfoAddDetailDto;
+import wizard.eVC.common.util.FTP;
 
 import java.io.IOException;
+import java.util.HashMap;
+
 import java.util.List;
 import java.util.Map;
 
@@ -97,10 +98,17 @@ public class InfoAddDetailController {
             //수정 안 했으면 삭제 여부 확인
             //빈칸이면 삭제로 판단
             if(infoAddDetailDto.attachFile.isEmpty()){
-                //파일 삭제 TODO
-                infoAddDetailDto.setAttachFile("");
-                infoAddDetailDto.setAttachPath("");
-                ftp.deleteFile(infoAddDetailDto.getDeleteAttachFile(),FTPPATH + "/" + infoID + "/");
+
+                if(infoAddDetailDto.deleteAttachFile.isEmpty()){
+                    infoAddDetailDto.setAttachFile("");
+                    infoAddDetailDto.setAttachPath("");
+                }else{
+                    //파일 삭제 TODO
+                    infoAddDetailDto.setAttachFile("");
+                    infoAddDetailDto.setAttachPath("");
+                    ftp.deleteFile(infoAddDetailDto.getDeleteAttachFile(),FTPPATH + "/" + infoID + "/");
+                }
+
             }
             else{
                 //수정 안 했으면 기존값 저장
